+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/_help.py
@@ -258,7 +258,7 @@ helps['iot device message send'] = """
     type: command
     short-summary: Send a device-to-cloud message.
     examples:
-        - name: Send a device-to-cloud message to an IoT Hub with a default message.
+        - name: Send a device-to-cloud message to an IoT Hub with a default message using amqp, mqtt or http.
           text: >
             az iot device message send --hub-name MyIotHub --device-id MyDevice
         - name: Send a device-to-cloud message to an IoT Hub with a custom message.
@@ -303,4 +303,40 @@ helps['iot device import'] = """
     type: command
     short-summary: Import, update, or delete device identities in the IoT hub identity registry from a blob.
     long-summary: For more information, see https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-identity-registry#import-and-export-device-identities.
+"""
+
+helps['iot device twin'] = """
+    type: group
+    short-summary: Manage device twins in your Azure IoT hub.
+"""
+
+helps['iot device twin show'] = """
+    type: command
+    short-summary: Output device twin definition.
+"""
+
+helps['iot device twin update'] = """
+    type: command
+    short-summary: Update device twin definition.
+"""
+
+helps['iot device method'] = """
+    type: command
+    short-summary: Invoke method on device.
+"""
+
+helps['iot device sas'] = """
+    type: command
+    short-summary: Generate a shared access signature token for the given device.
+"""
+
+helps['iot device simulate'] = """
+    type: command
+    short-summary: Simulate a device on your Azure IoT hub.
+    long-summary: In preview.
+"""
+
+helps['iot hub message send'] = """
+    type: command
+    short-summary: Send a cloud-to-device message.
 """