+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/custom.py
@@ -2,12 +2,16 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-# pylint: disable=no-self-use,no-member,line-too-long,too-few-public-methods
+# pylint: disable=no-self-use,no-member,line-too-long,too-few-public-methods,no-name-in-module
 
-from __future__ import print_function
 from os.path import exists
 from enum import Enum
+import json
+import time
+import uuid
+import six.moves
 from azure.cli.core.util import CLIError
+from azure.cli.core.util import read_file_content
 from azure.mgmt.iothub.models.iot_hub_client_enums import IotHubSku, AccessRights
 from azure.mgmt.iothub.models.iot_hub_description import IotHubDescription
 from azure.mgmt.iothub.models.iot_hub_sku_info import IotHubSkuInfo
@@ -17,6 +21,11 @@ from azure.cli.command_modules.iot.mgmt_iot_hub_device.lib.models.authentication
 from azure.cli.command_modules.iot.mgmt_iot_hub_device.lib.models.device_description import DeviceDescription
 from azure.cli.command_modules.iot.mgmt_iot_hub_device.lib.models.x509_thumbprint import X509Thumbprint
 from azure.cli.command_modules.iot.sas_token_auth import SasTokenAuthentication
+from azure.cli.command_modules.iot.iot_sdk.device_manager import DeviceManager
+from azure.cli.command_modules.iot.iot_sdk.utility import block_stdout
+from azure.cli.command_modules.iot.iot_sdk.utility import Default_Msg_Callbacks
+from iothub_service_client import IoTHubDeviceTwin, IoTHubDeviceMethod, IoTHubError, IoTHubMessaging, IoTHubMessage
+from iothub_client import IoTHubTransportProvider, IoTHubClientError
 from ._factory import resource_service_factory
 from ._utils import create_self_signed_certificate
 
@@ -381,3 +390,178 @@ def _convert_perms_to_access_rights(perm_list):
         'deviceconnect_registryread_registrywrite_serviceconnect': AccessRights.registry_read_registry_write_service_connect_device_connect
     }
     return access_rights_mapping[perm_key]
+
+
+# IoT SDK Extensions #
+# Twin Ops
+
+
+def iot_twin_show(client, device_id, hub_name):
+    c = iot_hub_show_connection_string(client, hub_name)
+    try:
+        return json.loads(IoTHubDeviceTwin(c['connectionString']).get_twin(device_id))
+    except IoTHubError as e:
+        raise CLIError(e)
+
+
+def iot_twin_update(client, device_id, hub_name, update_json):
+    c = iot_hub_show_connection_string(client, hub_name)
+    try:
+        # The SDK is looking for a raw json string
+        if exists(update_json):
+            update_json = str(read_file_content(update_json))
+        json.loads(update_json)
+        return json.loads(IoTHubDeviceTwin(c['connectionString']).update_twin(device_id, update_json))
+    except ValueError:
+        raise CLIError('Improperly formatted JSON!')
+    except IoTHubError as e:
+        raise CLIError(e)
+
+
+# Device Method Invoke
+
+def iot_device_method(client, device_id, hub_name, method_name, method_payload, timeout=60):
+    try:
+        c = iot_hub_show_connection_string(client, hub_name)
+        iothub_device_method = IoTHubDeviceMethod(c['connectionString'])
+        response = iothub_device_method.invoke(device_id, method_name, method_payload, timeout)
+        return {
+            'status': response.status,
+            'payload': response.payload
+        }
+    except IoTHubError as e:
+        raise CLIError(e)
+
+# Utility
+
+
+def iot_get_sas_token(client, hub_name, device_id, policy_name, duration=3600, resource_group_name=None):
+    resource_group_name = _ensure_resource_group_name(client, resource_group_name, hub_name)
+    base_url = '{0}.azure-devices.net'.format(hub_name)
+    uri = '{0}/devices/{1}'.format(base_url, device_id)
+    access_policy = iot_hub_policy_get(client, hub_name, policy_name, resource_group_name)
+    result = SasTokenAuthentication(uri, access_policy.key_name, access_policy.primary_key,
+                                    duration).generate_sas_token().replace('SharedAccessSignature ', '')
+    return {"SharedAccessSignature": result}
+
+# Messaging
+
+
+def iot_device_send_message_ext(client, device_id, hub_name, protocol='http', data='Ping from Azure CLI',
+                                resource_group_name=None, message_id=None, correlation_id=None, user_id=None):
+    try:
+        c = _get_single_device_connection_string(client, hub_name, device_id, resource_group_name, None)
+        protocol = _iot_sdk_device_process_protocol(protocol)
+        with block_stdout():
+            device = DeviceManager(c, protocol)
+            device.send_message(data, {'UserId': user_id} if user_id else None, message_id, correlation_id)
+    except IoTHubClientError as e:
+        raise CLIError(e)
+    except RuntimeError as f:
+        raise CLIError(f)
+
+
+def iot_hub_message_send(client, device_id, hub_name, message_id=str(uuid.uuid4()), correlation_id=None,
+                         data="Ping from Azure CLI", wait_feedback=False):
+    try:
+        c = iot_hub_show_connection_string(client, hub_name)
+        iothub_messaging = IoTHubMessaging(c['connectionString'])
+
+        message = IoTHubMessage(data)
+
+        # optional: assign ids
+        if correlation_id is not None:
+            message.correlation_id = correlation_id
+        if message_id is not None:
+            message.message_id = message_id
+
+        default = Default_Msg_Callbacks()
+
+        iothub_messaging.open(default.open_complete_callback, 0)
+
+        if wait_feedback:
+            iothub_messaging.set_feedback_message_callback(default.feedback_received_callback, 0)
+
+        iothub_messaging.send_async(device_id, message, default.send_complete_callback, 0)
+        time.sleep(2)
+
+        if wait_feedback:
+            wait_feedback_msg = "Waiting for message feedback, press any key to continue...\n\n"
+            six.print_('', flush=True)
+            six.moves.input(wait_feedback_msg)
+
+        iothub_messaging.close()
+    except IoTHubError as e:
+        raise CLIError(e)
+
+
+# Simulate Device
+
+class ProtocolType(Enum):
+    http = 'http'
+    amqp = 'amqp'
+    mqtt = 'mqtt'
+
+
+class SettleType(Enum):
+    complete = 'complete'
+    abandon = 'abandon'
+    reject = 'reject'
+
+
+def iot_simulate_device(client, device_id, hub_name, settle='complete', protocol='amqp', data="Ping from Azure CLI",
+                        message_count=5, message_interval=1, receive_count=None, file_path=None):
+    if message_count < 1:
+        raise CLIError("message_count must be > 0!")
+    if message_interval < 1:
+        raise CLIError("message_interval must be > 0!")
+
+    try:
+        protocol = _iot_sdk_device_process_protocol(protocol)
+        c = _get_single_device_connection_string(client, hub_name, device_id, None, None)
+        with block_stdout():
+            sim_client = DeviceManager(c, protocol)
+
+        if file_path:
+            sim_client.upload_file_to_blob(file_path)
+
+        if receive_count:
+            sim_client.configure_receive_settle(settle)
+
+        for message_counter in range(0, message_count):
+            print_context = "Sending message %s, via %s with %s sec delay" % (message_counter + 1,
+                                                                              protocol, message_interval)
+            sim_client.send_message(data, None, str(uuid.uuid4()), str(uuid.uuid4()), 0, print_context)
+            time.sleep(message_interval)
+
+        if receive_count:
+            while sim_client.received() < receive_count:
+                time.sleep(1)
+
+    except IoTHubClientError as e:
+        raise CLIError("Unexpected client error %s" % e)
+    except RuntimeError as f:
+        raise CLIError("Unexpected runtime error %s" % f)
+
+
+def _iot_sdk_device_process_protocol(protocol_string):
+    protocol = IoTHubTransportProvider.HTTP
+    if protocol_string == "http":
+        if hasattr(IoTHubTransportProvider, "HTTP"):
+            protocol = IoTHubTransportProvider.HTTP
+        else:
+            raise CLIError("HTTP protocol is not supported")
+
+    elif protocol_string == "amqp":
+        if hasattr(IoTHubTransportProvider, "AMQP"):
+            protocol = IoTHubTransportProvider.AMQP
+        else:
+            raise CLIError("AMQP protocol is not supported")
+
+    elif protocol_string == "mqtt":
+        if hasattr(IoTHubTransportProvider, "MQTT"):
+            protocol = IoTHubTransportProvider.MQTT
+        else:
+            raise CLIError("Error: MQTT protocol is not supported")
+
+    return protocol